requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.5
matplotlib>=3.3.4
scikit-learn>=0.24.1
tqdm>=4.60.0
Pillow>=8.2.0
pandas>=1.2.4
tensorboard>=2.5.0
PyQt5>=5.15.0 